requirements.txt
streamlit>=1.28.0
PyPDF2>=3.0.1
langchain>=0.1.0
langchain-community>=0.0.20
langchain-openai>=0.0.5
faiss-cpu==1.8.0
tiktoken>=0.5.2
python-dotenv>=1.0.0
openai>=1.6.1
sentence-transformers>=2.2.0
huggingface-hub>=0.16.0
transformers>=4.21.0
numpy==1.26.2 